+++ examples/room/room_server_main.cpp
@@ -42,6 +42,9 @@ int main(int argc, char* argv[]) {
     game::gateway::PushService push;
     game::room::RoomManager room_mgr;
     game::battle::BattleManager battle_mgr;
+    room_mgr.set_battle_active_query([&battle_mgr](const std::string& room_id) {
+        return battle_mgr.battle_started(room_id);
+    });
 
     game::room::RoomService room_svc(session_mgr, push, battle_mgr, room_mgr, metrics);
     room_svc.register_handlers(dispatcher);